feat: separate fastify app from its server listening mode
This separation, while seamingly artificial, helps to avoid
the Fastify app server in the global scope from the server.js
file.

It also aids in testing, where only the `appFramework` can be
used to fire up a listener bound to a port, or use Fastify's
own route injection capabilities (a la `fastify.inject()`)
in order to completely remove the need for an HTTP request
and a client to trigger one.